refactor: InputView, OutputView 세부 분리
