+++ src/main/kotlin/blackjack/view/ProgressOutputView.kt
@@ -4,7 +4,7 @@ import blackjack.model.Participant
 import blackjack.model.Participant.Dealer
 import blackjack.model.Participant.Player
 
-object OutputView {
+object ProgressOutputView {
     private const val MESSAGE_CARD_DISTRIBUTION = "\n%s와 %s에게 2장의 카드를 나누었습니다."
     private const val MESSAGE_DEALER_CARD_INFORMATION = "%s: %s"
     private const val MESSAGE_PARTICIPANT_CARD_INFORMATION = "%s 카드: %s"
